refactor: command 서비스와 query 서비스 분리
